refactor: Refatora o controller e o service do Ninja para reponder com ReponsesEntitys

+++ src/main/java/dev/lest/CadastroNinja/Ninjas/NinjaService.java
@@ -1,5 +1,7 @@
 package dev.lest.CadastroNinja.Ninjas;
 
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
@@ -19,16 +21,20 @@ public class NinjaService {
         this.repository = repository;
     }
 
-    public List<NinjaDTO> listarNinjas() {
+    public ResponseEntity<List<NinjaDTO>> listarNinjas() {
         List<NinjaModel> ninjaModel = repository.findAll();
-        return ninjaModel.stream()
-                .map(ninjaMapper::map)
-                .collect(Collectors.toList());
+        List<NinjaDTO> ninjaDTOList =  ninjaModel.stream()
+                                                .map(ninjaMapper::map)
+                                                .collect(Collectors.toList());
+        return ResponseEntity.ok(ninjaDTOList);
     }
 
-    public NinjaDTO ninjaPorId(long id) {
+    public ResponseEntity<?> ninjaPorId(long id) {
         Optional<NinjaModel> ninjaModel = repository.findById(id);
-        return ninjaModel.map(ninjaMapper::map).orElse(null);
+        if (ninjaModel.isPresent()) {
+            return ResponseEntity.ok(ninjaModel.map(ninjaMapper::map).orElse(null));
+        }
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body("O ninja de id: "+ id + " não foi encontrado!");
     }
 
     public NinjaDTO criarNinja(NinjaDTO ninja){
@@ -37,19 +43,24 @@ public class NinjaService {
         return ninjaMapper.map(ninjaModel);
     }
 
-    public void deletarNinjaPorId(long id) {
-        repository.deleteById(id);
+    public ResponseEntity<String> deletarNinjaPorId(long id) {
+        Optional<NinjaModel> ninjaExistente = repository.findById(id);
+        if (ninjaExistente.isPresent()) {
+            repository.deleteById(id);
+            return ResponseEntity.ok("Ninja do id: " + id + " foi deletado com sucesso!");
+        }
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body("O ninja de id: "+ id + " não foi encontrado!");
     }
 
-    public NinjaDTO alterarNinja(NinjaDTO ninja, long id){
+    public ResponseEntity<?> alterarNinja(NinjaDTO ninja, long id){
         Optional<NinjaModel> ninjaExistente = repository.findById(id);
         if (ninjaExistente.isPresent()) {
             NinjaModel ninjaAtualizado = ninjaMapper.map(ninja);
             ninjaAtualizado.setId(id);
             NinjaModel ninjaSalvo = repository.save(ninjaAtualizado);
-            return ninjaMapper.map(ninjaSalvo);
+            return ResponseEntity.ok(ninjaSalvo);
         }
-        return null;
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body("O ninja de id: "+ id + " não foi encontrado!");
     }
 
 }
